Combine 1.12.2 tests
