+++ src/test/kotlin/xyz/wagyourtail/unimined/test/integration/Minecraft1Point12Point2Test.kt
@@ -1,0 +1,29 @@
+package xyz.wagyourtail.unimined.test.integration
+
+import org.gradle.testkit.runner.TaskOutcome
+import org.gradle.testkit.runner.UnexpectedBuildFailure
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.MethodSource
+import xyz.wagyourtail.unimined.util.runTestProject
+
+class Minecraft1Point12Point2Test {
+    @ParameterizedTest
+    @MethodSource("xyz.wagyourtail.unimined.util.IntegrationTestUtils#versions")
+    fun testMinecraft1Point12Point2(gradleVersion: String) {
+        try {
+            val result = runTestProject("minecraft/1.12.2", gradleVersion)
+
+            try {
+                result.task(":build")?.outcome?.let {
+                    if (it != TaskOutcome.SUCCESS) throw Exception("build failed")
+                } ?: throw Exception("build failed")
+            } catch (e: Exception) {
+                println(result.output)
+                throw Exception(e)
+            }
+        } catch (e: UnexpectedBuildFailure) {
+            println(e)
+            throw Exception("build failed", e)
+        }
+    }
+}
