--- testing/1.12.2-Cleanroom/build.gradle
@@ -1,56 +1,0 @@
-plugins {
-    id 'java'
-    id 'xyz.wagyourtail.unimined' // version '1.0.0'
-}
-
-group 'com.example'
-version '1.0-SNAPSHOT'
-
-base {
-    archivesName = 'UniminedExampleMod'
-}
-
-// this is just here so we can test the outputs easier and clean between tests
-unimined.useGlobalCache = false
-
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
-
-repositories {
-    maven {
-        name = "wagyourtail releases"
-        url = "https://maven.wagyourtail.xyz/releases"
-    }
-    maven {
-        name = "sponge"
-        url = "https://repo.spongepowered.org/maven"
-    }
-    mavenCentral()
-}
-
-unimined.minecraft {
-    version project.minecraft_version
-
-    mappings {
-        mcp("stable", "39-1.12")
-    }
-
-    cleanroom {
-        loader "0.2.2-alpha"
-    }
-}
-
-dependencies {
-}
-
-processResources {
-    inputs.property "version", project.version
-
-    filesMatching("META-INF/mods.toml") {
-        expand "version": project.version
-    }
-
-    filesMatching("mcmod.info") {
-        expand "version": project.version
-    }
-}
