--- testing/1.12.2-Cleanroom/gradle.properties
@@ -1,5 +1,0 @@
-org.gradle.jvmargs = -Xmx2G
-
-minecraft_version = 1.12.2
-forge_version = 14.23.5.2860
-fabric_version = 0.14.20
