--- testing/1.12.2-Cleanroom/src/main/resources/META-INF/mods.toml
@@ -1,20 +1,0 @@
-modLoader="javafml"
-loaderVersion="[36,)"
-issueTrackerURL=""
-license = "MPL-2.0"
-
-[[mods]]
-modId = "modid"
-version = "${version}"
-displayName = "Example Mod"
-authors = "Me!"
-description = '''
-    This is an example description! Tell everyone what your mod is about!
-'''
-
-[[dependencies.modid]]
-modId="minecraft"
-mandatory=true
-versionRange="[1.8.9]"
-ordering="NONE"
-side="BOTH"
