--- testing/1.12.2-Cleanroom/src/main/resources/pack.mcmeta
@@ -1,6 +1,0 @@
-{
-  "pack": {
-    "description": "example resources",
-    "pack_format": 6
-  }
-}
