--- testing/1.12.2-Forge-Fabric/src/fabric/java/com/example/fabric/ExampleModFabric.java
@@ -1,13 +1,0 @@
-package com.example.fabric;
-
-import com.example.ExampleMod;
-import net.fabricmc.api.ModInitializer;
-
-public class ExampleModFabric implements ModInitializer {
-
-    @Override
-    public void onInitialize() {
-        ExampleMod.LOGGER.info("Hello from Fabric!");
-    }
-
-}
