--- testing/1.12.2-Forge-Fabric/src/forge/resources/mcmod.info
@@ -1,16 +1,0 @@
-[
-  {
-    "modid": "modid",
-    "name": "Example Mod",
-    "description": "This is an example description! Tell everyone what your mod is about!",
-    "version": "${version}",
-    "mcversion": "1.12.2",
-    "url": "",
-    "updateUrl": "",
-    "authorList": ["Me!"],
-    "credits": "",
-    "logoFile": "/assets/modid/icon.png",
-    "screenshots": [],
-    "dependencies": []
-  }
-]
