--- testing/1.12.2-Forge-Fabric/src/forge/resources/pack.mcmeta
@@ -1,6 +1,0 @@
-{
-  "pack": {
-    "description": "example resources",
-    "pack_format": 6
-  }
-}
