--- testing/1.12.2-Forge-Fabric/src/main/java/com/example/ExampleMod.java
@@ -1,9 +1,0 @@
-package com.example;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-public class ExampleMod {
-    public static final Logger LOGGER = LogManager.getLogger("modid");
-
-}
