+++ testing/minecraft/1.12.2/gradle.properties
@@ -1,0 +1,7 @@
+org.gradle.jvmargs = -Xmx2G
+
+minecraft_version = 1.12.2
+forge_version = 14.23.5.2860
+cleanroom_version = 0.2.2-alpha
+fabric_version = 0.16.0
+quilt_version = 0.26.3
