+++ testing/minecraft/1.12.2/src/forge/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+  "pack": {
+    "description": "Example mod resources",
+    "pack_format": 6
+  }
+}
