+++ testing/minecraft/1.12.2/src/legacyFabric/resources/fabric.mod.json
@@ -1,0 +1,25 @@
+{
+  "schemaVersion": 1,
+  "id": "modid",
+  "version": "${version}",
+  "name": "Example Mod",
+  "description": "This is an example description! Tell everyone what your mod is about!",
+  "authors": [
+    "Me!"
+  ],
+  "license": "MIT",
+  "environment": "*",
+  "entrypoints": {
+    "main": [
+		"com.example.example.mod.FabricExampleMod"
+    ]
+  },
+  "mixins": [
+  ],
+
+  "depends": {
+    "fabricloader": ">=0.16.0",
+    "minecraft": "~1.12.2",
+    "java": ">=8"
+  }
+}
