+++ testing/minecraft/1.12.2/src/legacyMain/java/com/example/example/mod/ExampleMod.java
@@ -1,0 +1,8 @@
+package com.example.example.mod;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+public class ExampleMod {
+    public static final Logger LOGGER = LogManager.getLogger("modid");
+}
