+++ testing/minecraft/1.12.2/src/mcpFabric/java/com/example/example/mod/FabricExampleMod.java
@@ -1,0 +1,10 @@
+package com.example.example.mod;
+
+import net.fabricmc.api.ModInitializer;
+
+public class FabricExampleMod implements ModInitializer {
+    @Override
+    public void onInitialize() {
+        ExampleMod.LOGGER.info("Hello from Fabric!");
+    }
+}
